added typescript; did some restructuring; added container component
