--- components/container/Container.jsx
@@ -1,17 +1,0 @@
-import Head from "next/head";
-import React from "react";
-
-function Container({ children, title, description }) {
-  return (
-    <>
-      <Head>
-        <title>{title}</title>
-        <meta name="description" content={description} />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main>{children}</main>
-    </>
-  );
-}
-
-export default Container;
