+++ components/info-bar/InfoBar.module.scss
@@ -1,0 +1,1 @@
+@use "../../styles/variables" as *;
