+++ package.json
@@ -16,5 +16,9 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sass": "^1.57.1"
+  },
+  "devDependencies": {
+    "@types/node": "18.11.18",
+    "@types/react": "18.0.26"
   }
 }
