+++ pages/_app.js
@@ -1,5 +1,5 @@
-import '../styles/globals.css'
+import "../styles/globals.scss";
 
 export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return <Component {...pageProps} />;
 }
