+++ pages/index.tsx
@@ -1,0 +1,12 @@
+import styles from "../styles/Home.module.scss";
+import Container from "../components/container";
+export default function Home() {
+  return (
+    <Container
+      title={"Home"}
+      description={"The landing page for a test restaurant site"}
+    >
+      <h1>Landing Page</h1>
+    </Container>
+  );
+}
